pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sploot-media-clustering"
version = "0.1.0"
description = "Media clustering service for Sploot"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.115.2",
    "uvicorn[standard]==0.30.6",
    "pydantic==2.9.2",
    "pydantic-settings==2.6.0",
    "redis==5.1.1",
    "numpy<2.0",
    "scikit-learn==1.5.2",
    "python-dotenv==1.0.1",
    "prometheus-client==0.21.0",
    "torch==2.1.2",
    "torchvision==0.16.2",
    "pillow==10.4.0",
    "timm==0.9.16",
    "aiohttp==3.10.5",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "fakeredis==2.21.3",
]

[tool.setuptools]
packages = ["sploot_media_clustering"]
package-dir = {"" = "src"}
